ai-bot/requirements.txt
# Matrix SDK for Python
matrix-nio==0.20.2

# HTTP client
aiohttp==3.8.6
requests==2.31.0

# LLM providers
openai==1.3.7
anthropic==0.7.8

# Web framework for application service
flask==3.0.0

# Environment management
python-dotenv==1.0.0

# JSON handling
jsonschema==4.19.2

# Async support
asyncio-throttle==1.0.2

# Logging
coloredlogs==15.0.1

# Date/time handling
python-dateutil==2.8.2

# Markdown processing
markdown==3.5.1

# Emoji support
emoji==2.8.0

# Configuration management
pyyaml==6.0.1